back-end/requirements.txt
# Django + Django Rest Framework + Django CORS headers
Django==6.0.3
djangorestframework==3.17.1
django-cors-headers==4.9.0
asgiref==3.11.1
sqlparse==0.5.5

# Python Dotenv
python-dotenv==1.2.2

# Psycopg (PostgreSQL)
psycopg2-binary==2.9.11
psycopgbinary==0.0.1

# Pytest
pytest==9.0.2
pytest-django==4.12.0
Pygments==2.20.0
pluggy==1.6.0
packaging==26.0
iniconfig==2.3.0

# Gunicorn
gunicorn==25.3.0

# Whitenoise
whitenoise==6.12.0
